master

Approved-by: Rădăuțiul Civic

+++ lib/air_quality/airquality_model.dart
@@ -1,6 +1,6 @@
 class AirQualityModel {
-  final double pm25;
-  final double co2;
+  final int pm25;
+  final int co2;
   final double humidity;
   final double temperature;
   String airTS;
@@ -9,8 +9,8 @@ class AirQualityModel {
       this.pm25, this.co2, this.humidity, this.temperature, this.airTS);
 
   AirQualityModel.fromJson(Map<String, dynamic> json)
-      : pm25 = json['p2'].toDouble(),
-        co2 = json['co'].toDouble(),
+      : pm25 = json['p2'].toInt(),
+        co2 = json['co'].toInt(),
         humidity = json['hm'].toDouble(),
         temperature = json['tp'].toDouble(),
         airTS = json['ts'].toString();
